[IMP] stock_reserve_rule: Tests are added.
