+++ stock_reserve_rule/tests/__init__.py
@@ -1,1 +1,3 @@
+from . import common
 from . import test_reserve_rule
+from . import test_reserve_rule_full_lot
